# Implements indexing for <container> elements as nested Solr Documents