--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -145,6 +145,9 @@ jobs:
       - store_artifacts:
           path: /tmp/test-results
           destination: test-results
+      - store_artifacts:
+          path: ~/pulfalight/coverage
+          destination: coverage
 
   coverage_report:
     working_directory: ~/pulfalight