--- a/app/services/container_id_minter_service.rb
+++ b/app/services/container_id_minter_service.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class ContainerIdMinterService < Ead2IdMinterService
+  NODE_NAME_REGEX = /^container$/.freeze
+end