--- a/lib/pulfalight/traject/ead2_container_config.rb
+++ b/lib/pulfalight/traject/ead2_container_config.rb
@@ -19,8 +19,32 @@ require Rails.root.join("lib", "pulfalight", "traject", "ead2_indexing")
 module ContainerIndexer
   self.class.include(Pulfalight::Ead2Indexing)
 
+  def hash_algorithm
+    Digest::SHA1
+  end
+
+  COMPONENT_NODE_NAME_REGEX = /^c\d{,2}$/.freeze
+
   def add_container_indexing_steps
     configure_before
+
+    to_field "id" do |record, accumulator|
+      # This needs to become a service class
+      ancestor_tree = record.ancestors.map do |ancestor|
+        if COMPONENT_NODE_NAME_REGEX.match?(ancestor.name)
+          ancestor_siblings = ancestor.parent.children.select { |n| n.name =~ COMPONENT_NODE_NAME_REGEX }
+          index = ancestor_siblings.index(ancestor)
+          "#{ancestor.name}#{index}"
+        else
+          ancestor.name
+        end
+      end
+
+      siblings = record.parent.children.select { |n| n.name =~ COMPONENT_NODE_NAME_REGEX }
+      current_index = siblings.index(record)
+      absolute_xpath = "#{[ancestor_tree.reverse, record.name].flatten.join('/')}#{current_index}"
+      accumulator << hash_algorithm.hexdigest(absolute_xpath).prepend("al_")
+    end
 
     to_field "type_ssim" do |record, accumulator|
       attribute = record.attribute("type")