--- a/lib/pulfalight/traject/ead2_container_config.rb
+++ b/lib/pulfalight/traject/ead2_container_config.rb
@@ -9,6 +9,7 @@ require "arclight/missing_id_strategy"
 extend TrajectPlus::Macros
 # rubocop:enable Style/MixinUsage
 require Rails.root.join("lib", "pulfalight", "traject", "ead2_indexing")
+require Rails.root.join("app", "services", "container_id_minter_service")
 
 # =============================
 # Each container child document
@@ -19,31 +20,13 @@ require Rails.root.join("lib", "pulfalight", "traject", "ead2_indexing")
 module ContainerIndexer
   self.class.include(Pulfalight::Ead2Indexing)
 
-  def hash_algorithm
-    Digest::SHA1
-  end
-
-  COMPONENT_NODE_NAME_REGEX = /^c\d{,2}$/.freeze
-
   def add_container_indexing_steps
     configure_before
 
     to_field "id" do |record, accumulator|
-      # This needs to become a service class
-      ancestor_tree = record.ancestors.map do |ancestor|
-        if COMPONENT_NODE_NAME_REGEX.match?(ancestor.name)
-          ancestor_siblings = ancestor.parent.children.select { |n| n.name =~ COMPONENT_NODE_NAME_REGEX }
-          index = ancestor_siblings.index(ancestor)
-          "#{ancestor.name}#{index}"
-        else
-          ancestor.name
-        end
-      end
-
-      siblings = record.parent.children.select { |n| n.name =~ COMPONENT_NODE_NAME_REGEX }
-      current_index = siblings.index(record)
-      absolute_xpath = "#{[ancestor_tree.reverse, record.name].flatten.join('/')}#{current_index}"
-      accumulator << hash_algorithm.hexdigest(absolute_xpath).prepend("al_")
+      minter = ContainerIdMinterService.new(node: record)
+      minted_id = minter.mint
+      accumulator << minted_id
     end
 
     to_field "type_ssim" do |record, accumulator|