Resolves #246