@jrgriffiniii 2020-04-24
The `rspec` test suite is now consuming 12 minutes on CircleCI - perhaps one of the test fixtures is the root cause of this problem.