@jrgriffiniii 2020-04-24
The coverage analysis is not accurate given that the `rspec` test suite is exceeding 10 minutes in length:

```
Too long with no output (exceeded 10m0s): context deadline exceeded 
```
(please see https://app.circleci.com/pipelines/github/pulibrary/pulfalight/432/workflows/1579d98a-9206-4ff5-8b52-4171c2e3106e/jobs/661)

For various reasons, the test suite needs to be simplified in order to resolve this.