@jrgriffiniii 2020-04-24
It seems as if running Traject twice, even on the same EAD file, triggers a massive delay in the tests being completed.