@jrgriffiniii 2020-04-24
Performance for this is being significantly increased merely by the introduction of `context` blocks. Testing this locally, I used one case with and without an unnecessary `context` block:

Without the `context`:
`Finished in 1 minute 47.55 seconds (files took 5.2 seconds to load)`
With the new `context`:
`Finished in 2 minutes 9.6 seconds (files took 6.42 seconds to load)`

Given that this is testing the Traject configuration, I can look to try and just reduce the number of these `context` blocks, and reduce the complexity of the EAD fixtures.