@jrgriffiniii 2020-04-24
The recursive approach introduced in https://github.com/pulibrary/pulfalight/pull/255 is what is triggering the delays. This will need to be adjusted.