@jrgriffiniii 2020-06-29
Unless this can be reconciled with ArchivesSpace EADs (which already have container IDs), this is no longer necessary.